MultiDevices - CMS : QC Defect # 140
The wrapped type "message/imdn+xml" is missing in OneToOnechatSession.

Upon receipt of S&F Chat notification from the Application Server, this wrapped type was missing in the SDP of the 200 OK response from the stack.
So IMDN notifications were not sent by the AS in the opened MSRP session.