CMS: bad intitial state for synchronaized GC
Synchronized GC is created in state STARTED (reason code UNSPECIFIED). The consequence is
that the GC is rejoined at the first next network reconnection even if the GC has
expired and no message is queued.
Synchronized GC should be created in the state ABORTED (reason code BY_INACTIVITY) to avoid such
behavior.